requirements.txt
python-dotenv
langchain==0.3.21
langchain_community==0.3.20
langchain-chroma
langsmith==0.3.17
streamlit==1.31.1
streamlit_chat==0.1.1
langchain_ollama
chromadb
tree-sitter==0.21.3
tree-sitter-languages
fastembed
PyGithub